+++ iox_data_generator/src/lib.rs
@@ -91,7 +91,7 @@ pub async fn generate<T: DataGenRng>(
     end_datetime: Option<i64>,
     execution_start_time: i64,
     continue_on: bool,
-    batch_size: u16,
+    batch_size: usize,
 ) -> Result<usize> {
     let seed = spec.base_seed.to_owned().unwrap_or_else(|| {
         let mut rng = rand::thread_rng();
